* Linux: Trivial compile fixes for Screen and the time glue...
* Screen and pnet for Matlab & Octave on Linux rebuilt.

==> All MEX files on all platforms are now complete and in sync.
==> Ready for beta release.



git-svn-id: https://psychtoolbox-3.googlecode.com/svn/trunk@1550 f5efbe48-b203-0410-9429-fc5e098e5a54

+++ PsychSourceGL/Source/Linux/Base/PsychTimeGlue.c
@@ -455,7 +455,7 @@ int PsychSetThreadPriority(psych_thread* threadhandle, int basePriority, int twe
 	struct sched_param sp;
 	pthread_t thread;
 
-	if ((NULL != threadhandle) && (0x1 != threadhandle)) {
+	if ((NULL != threadhandle) && (0x1 != (int) threadhandle)) {
 		// Retrieve thread handle of thread to change:
 		thread = *threadhandle;
 	}
